# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Snap is an open telemetry framework designed to simplify the collection,\nprocessing and publishing of system data through a single API.\n\nThis should demonstrate all the possible comment annotations\nthat are available to turn go code into a fully compliant swagger 2.0 spec",
    "title": "Snap Restful API.",
    "termsOfService": "there are no TOS at this moment, use at your own risk we take no responsibility",
    "contact": {
      "name": "Snap Maintainer",
      "url": "http://snap.io",
      "email": "snap@intel.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/"
    },
    "version": "1.0.0"
  },
  "host": "127.0.0.1:8181",
  "basePath": "/v2",
  "paths": {
...
    "/plugins/{ptype}/{pname}/{pversion}/config": {
      "get": {
        "description": "Any other type results in error.",
        "consumes": [
          "application/json",
          "application/x-protobuf"
        ],
        "produces": [
          "application/json",
          "application/x-protobuf"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "lists the config of a giving plugin. The allowed plugin types are collector, processor, and publisher.",
        "operationId": "getPluginConfigItem",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "PName",
            "name": "pname",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "PVersion",
            "name": "pversion",
            "in": "path",
            "required": true
          },
          {
            "enum": [
              "collector",
              " processor",
              " publisher"
            ],
            "type": "string",
            "x-go-name": "PType",
            "name": "ptype",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/PluginConfigResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      },
      "put": {
        "consumes": [
          "application/json",
          "application/x-protobuf"
        ],
        "produces": [
          "application/json",
          "application/x-protobuf"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "updates the config of a giving plugin. A wrong plugin type or non-numeric plugin version results in error.",
        "operationId": "setPluginConfigItem",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Config",
            "name": "config",
            "in": "formData"
          },
          {
            "type": "string",
            "x-go-name": "PName",
            "name": "pname",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "PVersion",
            "name": "pversion",
            "in": "path",
            "required": true
          },
          {
            "enum": [
              "collector",
              " processor",
              " publisher"
            ],
            "type": "string",
            "x-go-name": "PType",
            "name": "ptype",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/PluginConfigResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      },
      "delete": {
        "description": "At lease one config key is required for this operation. An error occurs for any bad request.",
        "consumes": [
          "application/json",
          "application/x-protobuf"
        ],
        "produces": [
          "application/json",
          "application/x-protobuf"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "deletes the config of a giving plugin. Note that that to be removed config items are a slice of config keys.",
        "operationId": "deletePluginConfigItem",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "PName",
            "name": "pname",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "PVersion",
            "name": "pversion",
            "in": "path",
            "required": true
          },
          {
            "enum": [
              "collector",
              " processor",
              " publisher"
            ],
            "type": "string",
            "x-go-name": "PType",
            "name": "ptype",
            "in": "path",
            "required": true
          },
          {
            "type": "array",
            "items": {
              "type": "string"
            },
            "x-go-name": "Config",
            "name": "config",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/PluginConfigResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ConfigDataNode": {
      "description": "Represents a set of configuration data",
      "type": "object",
      "x-go-package": "github.com/intelsdi-x/snap/core/cdata"
    },
...
  },
  "responses": {
    "ErrorResponse": {
      "description": "Error unsuccessful generic response to a failed API call",
      "schema": {
        "type": "object",
        "additionalProperties": {
          "type": "string"
        }
      }
    },
...
    "PluginConfigResponse": {
      "description": "PluginConfigItem represents cdata.ConfigDataNode which implements it's own UnmarshalJSON.",
      "schema": {
        "$ref": "#/definitions/ConfigDataNode"
      }
    },
...
  }
}